Define code project changeset.
This commit also adds additional change types, such as replacing text
within a file and appending an entry to an INI file. The changeset
currently accepts vectors of input for API references and config files,
but we may wish to change this to accept a functor or to take advantage
of the `project_root` path that we are already consuming.