Add regex support for replace in file.
This makes the replacement operation more useful, albeit at the cost of
some efficiency. It is not obvious how to store a pattern field in a
struct, which makes it difficult to compile the regex while creating
the changeset. It may be worth exploring this further, as we can then
eliminate regex failures during engine error handling.