Implement backup mechanism.
Before any file is changed, its content is backed up into a backup
folder (`.renom/backup` in the typical case). The backed up file is
named using the SHA256 hash of the content, similar to the Git style.
This allows us to avoid naming conflicts/overwriting a backup file when
the filename is the same but content differs.