Give user choice to rename target files.
Previously, we assumed that target files would have the exact same name
as the project. However, this is not always the case. For example, the
Lyra game has project name LyraStarterGame but targets LyraGame and
LyraEditor. This commit identifies each of the relevant targets without
assuming what they will be in advance and gives the user a choice on
how to rename each.