Add support for multiple modules.
The previous setup assumed that there was only one module and that it
had the same name as the project. This assumption does not hold true
in many cases (e.g, Lyra). This update adds support for multiple
modules and prompts the user for a specific name for each rather than
using the project name.