Move API reference search into module changeset.
The API macros are only used within a given module, so we only need to
search within the module-specific folder to find references to the
macro. In addition, the original implementation only searched for
references in a folder that matched the project name; this prevented it
from finding files for modules that did not match the project name and
also prevented renaming multiple modules since at most references for
one module would be found.