refactor: refine mod implementation replace logic
This commit refines the mod implementation replace logic, which
originally did a full replace of old to new on the content of the
implementation macro. This causes issues when the module class includes
the old name, since we are not renaming that class. The updated logic
targets only the module name (and game name in case of a primary game
module). It is still not perfect but it is sufficient for most cases.