feat: remove target rename from project rename workflow
The original project rename workflow also detected targets and tried to
rename those as well. However, the rename support for targets was
limited. In addition, it was out of place because the original
assumption, that target files would have predictable names based on the
project name, proved false, and because a project rename does not
require renaming target files.