feat: rename cross-target references
It is possible for a project to have multiple build targets and for one
target to depend on another target. It is not clear how common this is,
but this is used in the Lyra starter project. When a project target is
renamed, references to that target must be updated in other targets as
well.