feat: add redirect support for plugin rename
Though code references do not change in a plugin rename, content
references will change as the root package element for plugin content is
the name of the plugin. This commit adds support to the rename plugin
workflow for redirects (adding a new one and updating existing ones) to
prevent breakage on rename.